koa 项目，做book路由的子路由hot_book，创建hot_book数据库，它只是个逻辑表业务表，不会存储图书的所有数据。图书的所有数据是另一个基础表。同时为了添加数据方便，做了对hot_book的图书的点赞和取消赞的路由，临时的。为了做这两个路由，在迁移对classic的点赞的逻辑的时候，发现其中的likeValidator错了，没有传递参数，一直校验的是登录方式，而不是点赞的数据的校验。修改了。同时修改了后面生成favor数据的方法，避免跟以前的数据库一致性的那部分逻辑的冲突。开始正式写hot_book的获取方法。

+++ koa/app/model/favor.js
@@ -1,4 +1,4 @@
-import {Model,  Sequelize} from "sequelize";
+import {Model, Sequelize} from "sequelize";
 import sequelize from "../../core/db";
 import {DislikeError, LikeError, NotFound} from "../../core/http-exception";
 // import Flow from "./flow";
@@ -17,11 +17,13 @@ class Favor extends Model {
         const favor = await Favor.findOne({
             where: {art_id, type, uid}
         });
+        console.log(art_id, type, uid, favor, "ff");
         if (favor) {
             throw new LikeError();
         }
 
         //Sequelize 的事务处理：https://github.com/demopark/sequelize-docs-Zh-CN/blob/master/transactions.md
+        if (type === 400) return (await Favor.create({art_id, uid, type}));
         return sequelize.transaction(async t => {
             //各个表的操作：
             //计入 用户跟art的关系！
@@ -38,11 +40,15 @@ class Favor extends Model {
         const favor = await Favor.findOne({
             where: {art_id, type, uid}
         });
+        console.log(art_id, type, uid, favor, "ff");
+
         if (!favor) {
             throw new DislikeError();
         }
 
         //Sequelize 的事务处理：https://github.com/demopark/sequelize-docs-Zh-CN/blob/master/transactions.md
+        if (type === 400) return (await favor.destroy());
+
         return sequelize.transaction(async t => {
             //各个表的操作：
             await favor.destroy({
@@ -67,12 +73,13 @@ class Favor extends Model {
         );
         // console.log(favor, uid);
         if (!favor.length) {
-            throw new NotFound("您还没有添加喜欢哦！",201);
+            throw new NotFound("您还没有添加喜欢哦！", 201);
         }
         const getIds = {
             100: [],
             200: [],
-            300: []
+            300: [],
+            400: []
         };
         favor.map(
             vals => {
